+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - Add `NewResourceDetectorWithOptions` and the `WithAWSLogger` option to `go.opentelemetry.io/contrib/detectors/aws/ec2/v2`, allowing a custom AWS SDK `logging.Logger` to be supplied to the EC2 resource detector. (#9132)
+- Add `WithRedactedQueryParams` to `go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp`, redacting the values of the given query parameters in the `url.full` span attribute recorded by `Transport`. (#9643)
 
 ### Changed